+++ src/port/posix/impl/udp_socket_impl.hpp
@@ -51,7 +51,7 @@ udp<Endpoint, Flags>::
 send(const void* buffer, std::size_t buffer_len, endpoint& ep, CoAP::Error& ec) noexcept
 {
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__) || defined(__NT__)
-	int sent = ::sendto(socket_, static_cast<const char*>(buffer), buffer_len, 0,
+	int sent = ::sendto(socket_, static_cast<const char*>(buffer), static_cast<int>(buffer_len), 0,
 				reinterpret_cast<struct sockaddr const*>(ep.native()),
 				sizeof(typename endpoint::native_type));
 #else /* defined(WIN32) || defined(_WIN32) || defined(__WIN32__) || defined(__NT__) */
@@ -77,7 +77,7 @@ receive(void* buffer, std::size_t buffer_len, endpoint& ep, CoAP::Error& ec) noe
 #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__) || defined(__NT__)
 	int addr_len = sizeof(struct sockaddr);
 	int recv = ::recvfrom(socket_,
-			static_cast<char*>(buffer), buffer_len, 0,
+			static_cast<char*>(buffer), static_cast<int>(buffer_len), 0,
 			reinterpret_cast<struct sockaddr*>(ep.native()), &addr_len);
 #else /* #if defined(WIN32) || defined(_WIN32) || defined(__WIN32__) || defined(__NT__) */
 	unsigned addr_len = sizeof(struct sockaddr);
